Dan commit
changes made
- added the extra 3 tab and frame for new ran strategies
i.e resultdisplay runs the default from the trading engine that mike
wrote
- strategyselected is essentially the same as resultdipaly with parts
changed such as removed strategy menu, and a possible differetn strategy
implementation to be used.
- transfered the lines printing from console to jlabel
- while doing this part, i found that it was slightly weird the way the
program
runs the strategy and updates the jlabel/ jgraphs etc.  when you run
strategy,
the strategy values and stuff are calculated before the labels are made.
this meant that i had to update the labels after creating the frame.
youll see several example in the code.
- the bit where it says IMPLEMENT STRATEGY in resultdisplay.java is
where u   put the matching and evaluation with the strategy that armin
wrote. the newrunstrategy in mainmenu is to run the default matching
mike did.
- graph creation and updated dataset value examples in is graphpanel()
- linegraph.java doesnt use the createfakedataset() so it is commented
out
- to implement the table changes, unless you somehow make the strategy
that was ran carry over its info to the frame with a public global in
resultdiplay/selectedstrategy. otherwise you might need to do it the
same way as how the jlabels were done to for the total lines bit
- fixed column names and fake data
- combined buy and sell table
- unfortunately gui still looks bland, took too long figuring out how to
update the labels after the strategy. but it has enough info to get by